Initial Commit of Hidouki
This is the first commit of the Hidouki web framework. It includes
a minimal async http server that unconditionally returns a 200
status code upon being connected to by a client. This also sets
up guidelines such as which companies we won't work with, Code of
Conduct and Licensing the code with The License Zero Prosperity License.

Signed-off-by: Michael Gattozzi <mgattozzi@gmail.com>