Add better error handling and response bodies
Before if the server encountered an error it would just completely shut
the entire server down. This is not desireable at all. Instead we now
will handle the error locally to keep the server running instead. This
means we'll respond with what caused the error to the user rather than
just letting it keep the connection hanging or with no response.