Add asynchronous route handling to Hidouki
If you want a webframework you need routing and since this is an async
framewosk we want async functions. However, there's one big problem.
When you write `async fn my_func() -> T` it desugars to `fn my_func() ->
impl Future<Output=T>`. Traits cannot use `impl Trait` anywhere. While
in a sync based environment we might use a regular `fn` pointer we can't
here. We'd have to Box the function somehow. This is bad for a few
reasons:

- The API becomes ergonomic and becomes something the callee will have
  to do when passing in the routes
- Any kind of nice syntactic proc macro sugar is impossible to do now
- We take performance hits due to the indirection, despite using named
  static functions

What are we to do then? Luckily for us `async_std` contains a way to
spawn async tasks that gives us a type that can be awaited that is not
`impl Future`. We wrap the actual async function call in the spawn
function inside of a sync function in a trait. This way we still have
the performance of an fn pointer, an easily procmacroable API, async
routing, and an ergonomic API. We get to have our cake and eat it too so
to speak. From here we can clean up the code, add TLS support, and
release an initial bare bones implementation for people to use.