Add a router macro to simplify setting up hidouki
While we could have people implement the Router trait manually it's
cumbersome, error prone, and ugly. We want people to have a nice
experience using this framework. Using a procmacro to declare the name
and the method used for the path over an async function will automatically
generate the required code for the Route impl so that it looks like they
are passing in a function into the vec, not a type. The little bit of
magic cleaned up the example, but we provide the expanded verion as
a comment so that people can see just how that code will expand.